AI/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pillow>=10.0.0
python-multipart==0.0.9

# YOLOv8/11 + 의존성
ultralytics==8.3.0
opencv-python-headless>=4.8.0

--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124
timm>=1.0.9
httpx>=0.27.0
paho-mqtt>=1.6.1
pydantic-settings>=2.6.0
